Rework: authorized_keys
authorized_keys now need to be copied into one file per user.

The default behaviour is that all keys in this file will be used exclusivly.
This can be changed by a new variable: wtd_users_default_auth_key_exclusive
When set to false it will not delete any keys already in the authorized_keys.